warning fixes, cloaks for vehicles, city map improved
- vehicle cloak now functioning
- cloak in battle functioning as per vanilla
- city map has colored icons for friendly/neutral/enemy craft
- strategy map now shows movement destination for all friendlies

+++ game/state/battle/battleunitanimationpack.cpp
@@ -379,7 +379,7 @@ void BattleUnitAnimationPack::drawUnit(
 			case AnimationEntry::Frame::UnitImagePart::Body:
 				if (!body)
 					continue;
-				if (body->images.size() <= b->index)
+				if ((int)body->images.size() <= b->index)
 				{
 					LogError("Missing image: body %s size %u index %d", body.id,
 					         body->images.size(), b->index);
@@ -391,7 +391,7 @@ void BattleUnitAnimationPack::drawUnit(
 			case AnimationEntry::Frame::UnitImagePart::Legs:
 				if (!legs)
 					continue;
-				if (legs->images.size() <= b->index)
+				if ((int)legs->images.size() <= b->index)
 				{
 					LogError("Missing image: legs %s size %u index %d", legs.id,
 					         legs->images.size(), b->index);
@@ -403,7 +403,7 @@ void BattleUnitAnimationPack::drawUnit(
 			case AnimationEntry::Frame::UnitImagePart::Helmet:
 				if (!helmet)
 					continue;
-				if (helmet->images.size() <= b->index)
+				if ((int)helmet->images.size() <= b->index)
 				{
 					LogError("Missing image: helmet %s size %u index %d", helmet.id,
 					         helmet->images.size(), b->index);
@@ -415,7 +415,7 @@ void BattleUnitAnimationPack::drawUnit(
 			case AnimationEntry::Frame::UnitImagePart::LeftArm:
 				if (!leftHand)
 					continue;
-				if (leftHand->images.size() <= b->index)
+				if ((int)leftHand->images.size() <= b->index)
 				{
 					LogError("Missing image: leftHand %s size %u index %d", leftHand.id,
 					         leftHand->images.size(), b->index);
@@ -427,7 +427,7 @@ void BattleUnitAnimationPack::drawUnit(
 			case AnimationEntry::Frame::UnitImagePart::RightArm:
 				if (!rightHand)
 					continue;
-				if (rightHand->images.size() <= b->index)
+				if ((int)rightHand->images.size() <= b->index)
 				{
 					LogError("Missing image: rightHand %s size %u index %d", rightHand.id,
 					         rightHand->images.size(), b->index);
@@ -439,7 +439,7 @@ void BattleUnitAnimationPack::drawUnit(
 			case AnimationEntry::Frame::UnitImagePart::Weapon:
 				if (!heldItem)
 					continue;
-				if (heldItem->held_image_pack->images.size() <= b->index)
+				if ((int)heldItem->held_image_pack->images.size() <= b->index)
 				{
 					LogError("Missing image: heldItem %s pack %s size %u index %d", heldItem.id,
 					         heldItem->held_image_pack.id, heldItem->held_image_pack->images.size(),
